리뷰를 가져올 때 특정 상황에서 앱이 팅기는 현상 수정

+++ FrontEnd/app/src/main/java/com/example/deliciousfood/pages/RestaurantActivity.java
@@ -149,7 +149,7 @@ public class RestaurantActivity extends ParentActivity {
                     if (reviewSearchResponseDTO != null) {
 
                         reviewSearchResponseDTO.getResult().forEach(review -> {
-                            reviewModels.add(new ReviewModel(review.getNickname(), Float.parseFloat(review.getScore()), review.getContent()));
+                            reviewModels.add(new ReviewModel(review.getNickname(), Integer.parseInt(review.getScore()), review.getContent()));
                         });
 
                         reviewAdapter.notifyDataSetChanged();
